Database Manager Created, backend integrated without swift data.
