--- SwiftBuddiesIOS/Targets/MapModule/Sources/MapView/Scenes/MainMapView/Manager/GetManager.swift
@@ -1,12 +1,0 @@
-//
-//  GetManager.swift
-//  SwiftBuddiesIOS
-//
-//  Created by Oğuzhan Abuhanoğlu on 10.11.2024.
-//
-
-import Foundation
-import BuddiesNetwork
-import Network
-
-class
